refactor!(crate): create request and item modules
This commit is a big refactor of the crate. It creates two new modules
`request` and `item` and moves the code from `src/requests.rs` and
`src/item.rs` into them. It also moves the `filter.rs` and `sort.rs`
files into the `request` module. 'item/attribute.rs' is a new file
that contains the `Attribute` struct and its implementation.

This commit serves as an improvement of the crate's structure.

BREAKING CHANGE: To call the 'Attribute' emun and its variants, you
must now use the 'item::Attribute' path (item::BookAttribute).